Smart-Library/pyproject.toml
[tool.poetry]
name = "smart library"
version = "0.1.0"
description = ""
authors = ["Yasser AlRuwaiti <Yasser.alruwaiti@pwc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
pyproject = "^1.3.1"
pre-commit = "^3.7.1"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
langchain-community = "^0.2.10"
langchain = "^0.2.11"
langchain-chroma = "^0.1.2"
sentence-transformers = "^3.0.1"
langchain-huggingface = "^0.0.3"
huggingface-hub = "^0.24.1"
openai = "^1.37.0"
tiktoken = "^0.7.0"
langchain-openai = "^0.1.17"
chromadb = "^0.5.5"
langchain-ollama = "^0.1.0"
ollama = "^0.3.0"
sseclient = "^0.0.27"
langchain-experimental = "^0.0.63"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
